test: test ThresholdCheckerMixin and WeightsDownloaderMixin separately
